requirements.txt
vector-quantize-pytorch
pytest
hydra-core
omegaconf